add Editable component
The Editable component is a simple component that, in essence, is an editable span. The user can click on the text to have it replaced by an input element containing that span's text. Upon losing focus, the input field will be replaced by the original span. Upon the enter key being released, the input field will be replaced again by a span, but the span's text will reflect the new value input by the user.