change editable component
Make several changes to the editable component. There are new inputs: inputRejectingFunctions, maxLength, plainTextClass, plainTextStyle, editClass, and editStyle. inputRejectingFunctions is an array of objects containing two properties. The first is a function mapping a string to a boolean. This function is meant to a be a validation function, that returns true if a new value input by the user satisfies the validation criteria, and false if it does not. The second property is an object. This object will be one of the elements emitted by the onInputRejected output if the new input fails this validation criteria. maxLength is the maxinum number of characters the input can hold at any given time. plainTextClass is a string that will be inserted into the class atribute of the plain text element of the editable. plainTextStyle is a string that will be applied as styling for the plain text element. editClass and editStyle are the counterparts and will be applied to the input element.
These changes are meant to make the editable component more robust. Components using editables now have control over how the component looks in both modes (plain text and editing). The input-rejecting functions can be used to prevent the value of the editable from being set to undesired values (rather than leaving this for future validation).